backend/app/requirements.txt
# ---------------- CORE FRAMEWORK ----------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ---------------- DATA & FINANCE ----------------
pandas
yfinance
requests
vaderSentiment

# ---------------- ASYNC & HTTP ----------------
httpx>=0.26.0
aiofiles>=23.2.1

# ---------------- ENVIRONMENT / CONFIG ----------------
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ---------------- DATABASE ----------------
motor

# ---------------- CACHE ----------------
redis>=5.0.1

# ---------------- TESTING ----------------
pytest>=7.4.3
pytest-asyncio>=0.23.3

# ---------------- FRONTEND ACCESS (CORS / MIDDLEWARE) ----------------
starlette>=0.35.0

# ---------------- AI AGENTS (LANGGRAPH) ----------------
langgraph>=0.0.62
langchain>=0.1.6
langchain-core>=0.1.23
langchain-community>=0.0.20

# ---------------- ADDITIONAL DEPENDENCIES ----------------
# These are often needed but may not be explicit
typing-extensions>=4.8.0
annotated-types>=0.6.0
pydantic-core>=2.14.0
anyio>=3.7.1
sniffio>=1.3.0
idna>=3.6
certifi>=2023.11.17